Extract studio components
